// astro-infoportal/src/Models/Generated/SubsidyPageViewModel.ts
/**
 * This is a TypeGen auto-generated file.
 * Any changes made to this file can be lost when this file is regenerated.
 */

import { IReactProps } from "./IReactProps";
import { IReactComponentProps } from "./IReactComponentProps";
import { BaseReactComponentProps } from "./BaseReactComponentProps";
import { RichTextAreaProps } from "./RichTextAreaProps";
import { TimelineItemBlockViewModel } from "./TimelineItemBlockViewModel";
import { BreadcrumbViewModel } from "./BreadcrumbViewModel";
import { ContentAreaProps } from "./ContentAreaProps";

export interface SubsidyPageViewModel extends IReactProps, IReactComponentProps, BaseReactComponentProps {
  pageName?: string;
  mainIntro?: string;
  mainBody?: RichTextAreaProps;
  timeline: TimelineItemBlockViewModel[];
  breadcrumb: BreadcrumbViewModel;
  lastUpdatedDateText?: string;
  lastUpdatedDateString?: string;
  bottomContentArea?: ContentAreaProps;
  commonBottomArea?: ContentAreaProps;
  agency?: string;
  coordinator?: string;
  industryConnection?: string;
  purposeConnection?: string;
}